refactor: Move elastic data handling to apartment app
